pyproject.toml
[project]
name = "tensorrt-llm-model-builder"
version = "0.1.0"
description = "TensorRT LLM Model Builder"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-storage>=3.1.0",
    "huggingface-hub>=0.31.1",
    "loguru>=0.7.3",
    "result>=0.17.0",
    "ty>=0.0.0a8",
]

[dependency-groups]
ci = [
    "mypy>=1.15.0",
    "ruff>=0.11.8",
]


[tool.mypy]
ignore_missing_imports = true